requirements.txt
streamlit==1.31.1
qdrant-client==1.7.0
python-dotenv==1.0.0
openai==1.13.3
fastembed==0.2.2
requests==2.31.0
beautifulsoup4==4.12.2 